Use entry_points rather than scripts in setup.py
This commit updates setup.py to use entry_point keyword rather
than scripts. This is because on Windows the former approach
resulted in incorect scripts. spinel-cli.py has been
renamed to spinel_cli.py as hyphen isn't allowed.
